Reload jobs after locking
This fixes a bug where failed jobs would never be unlocked.

I also marked the methods as protected, because it isn't safe to call
rails4's new changes_applied method if any other attributes have
changed. The one usage of the method is safe in this regard, since it's
a freshly-loaded instance.

Change-Id: I59ae293eea1f45121e897bcde7b2e2aef497f3e7