Add lifecycle event for error'd jobs
closes CNVS-19509

Canvas would like to be able to create an error report for
each job failure, so this lifecycle event lets it happen.

Also adds a spec for worker that runs regardless of backend (because
it's not actually adding jobs to a queue or anything, it's exercising
worker as a standalone component).

Finally, add a little documentation for hooking into lifecycle events,
and for running tests.

TEST PLAN:
 - setup something to run in a callback on ":error"
 - queue a job that will fail
 - confirm your callback gets run

Change-Id: Ie2272c04b3294dbd116e1eb5e40dbc6463324634