expires_at option
for time-sensitive jobs, enqueue with :expires_at => datetime
to set an expiration time. if the job is dequeued after this time,
it will immediately and permanently fail.

this is useful when a job is unnecessary or even harmful if run
too long after it was enqueued.

use a worker that responds to :on_permanent_failure to get
notified that a job has expired, or enqueue with
:on_permanent_failure => :some_method, for example:

widget.send_later_enqueue_args(:frob,
                               :expires_at => 1.day.from_now,
                               :on_permanent_failure => :log_error)

refs CNVS-25113

Change-Id: Ie2e21d81a0d40adffde0ddb3b7f6cea44c724826
Reviewed-on: https://gerrit.instructure.com/67744
Tested-by: Jenkins
Reviewed-by: James Williams  <jamesw@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>