change WorkQueue args to the full config
refs NTRS-651

Some extensions add their own arguments to the worker config that affect how
the queue is queried, so we want to pass the full config to the parent process.

We'll also want this for implementing coalesced job popping, where multiple
requests with the exact same worker config get batched into one queue query.

Change-Id: Ib8e9b9fde9107565f8f4660c4fde930c67ecc933
Reviewed-on: https://gerrit.instructure.com/75438
Tested-by: Jenkins
Reviewed-by: Tyler Pickett <tpickett@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>