Make socket path configurable, and move the default
Doing so will allow us to move the socket out of a path that
automatically gets cleaned up. Also, moving the default will make it
largely unnecessary to even move it around for our purposes.

Also, since we're adding additional configuration options for the
parent process queue option we've moved all of that config into a
parent_process hash in the yml file.

Refs: CNVS-36134

Change-Id: I649a5e9b9e00d7886d0846269837fa6c510b6e72
Reviewed-on: https://gerrit.instructure.com/109003
Tested-by: Jenkins
Reviewed-by: Simon Williams <simon@instructure.com>
Reviewed-by: Tyler Pickett <tpickett@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>