Stop clients hanging out waiting for a server that is gone
Fixes: CNVS-36134

There are a couple options to handle this situation; we could reach for
non-blocking reads but, then we would have to handle partial reads.
Another option would be to use Timeout.timeout but that would interrupt
a valid read that started just a few microseconds before the timeout
expired, so that's no good. This solution makes us wait until the socket
is ready but bails if we exceed our timeout.

Change-Id: I7078dd23663a162ce92439ab9d2970dac0e7d41a
Reviewed-on: https://gerrit.instructure.com/109436
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins