fix constant connection drops
don't use receive timeouts. we expect receives to block for possibly a
very long time until a job is available.

also, if we couldn't deliver a job to a client, unlock that job so
that it doesn't end up orphaned

finally, gracefully handle a client shutting down (jobs recycle
periodically)

Change-Id: I795874ca46a7ccc46dfe16a116050ebaecb407f0
Reviewed-on: https://gerrit.instructure.com/109926
Tested-by: Jenkins
Reviewed-by: Tyler Pickett <tpickett@instructure.com>