Fix accidental locking of too many jobs
The UPDATE statement that locked jobs was originally LIMITed in a
subquery that was then joined with the table being UPDATEd, but this
allowed the query planner to (sometimes) side-step the LIMIT, causing
too many jobs to become locked (but not returned).

Instead, we now use a CTE to force the subquery to be materialized
before joining to the table being UPDATEd. This ensures we don't lock
too many jobs.

For more details, see:
 * https://dba.stackexchange.com/a/69497/55285
 * https://github.com/feikesteenbergen/demos/blob/b7ecee8b2a79bf04cbcd74972e6bfb81903aee5d/bugs/update_limit_bug.txt

Fixes TS-3444

Change-Id: I2252723c8a158eeee041815ed781eff568d3669b
Reviewed-on: https://gerrit.instructure.com/137272
Tested-by: Jenkins
Reviewed-by: Tyler Pickett <tpickett@instructure.com>