Plumb SystemExit into on_failure callback, refs ARC-3362
During arc deploys SystemExit events results in any interrupted jobs to be
dropped on the floor if they are not configured with retries. This
provides at-most-once semantics when our jobs would prefer at-least-once
semantics.

To allow for more customization at the job level we'll capture the
return value from the on_failure callback and use that return value to
customize rescheduling behavior.

Change-Id: I75a772d47c219801c8ed3d5fe0c664e31f0a3a72
Reviewed-on: https://gerrit.instructure.com/150744
Reviewed-by: Michael Ziwisky <mziwisky@instructure.com>
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>